requirements.txt
aiohttp==3.9.1
numpy==1.26.3
PyYAML==6.0.1
colorama==0.4.6
prometheus-client==0.19.0
pydantic>=2.0.0
psutil>=5.9.0
